Key Pair Definition and a couple of miner edits
I suggested extending the definition of a keypair to emphasize the relationship they have together.  I also have suggested a couple minor edits.

I particularly appreciated your example of the multiplication of prime numbers and how difficult it would be to guess which prime numbers were used.  

Thanks for open-sourcing this book and giving the community the chance to contribute!